requirements.txt
django~=4.2.5
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
drf-yasg~=1.21.7